taxi: extract e-optimizer QSim module
